SEOArticleWriter/requirements.txt
python-dotenv>=1.0.0
requests>=2.0.0
openai>=0.27.0
langchain>=0.0.300
